V2RaycSpider1125/requirements.txt
tqdm
pyyaml
easygui
paramiko
pyperclip
requests
bs4
beautifulsoup4
schedule
gevent
flask
setuptools
pytz
redis
pillow
selenium
psutil
environs
loguru